Adds a log in flow and re-tools app coordinator
Adds a log in flow that represents the ability to go to a different root
flow for authenticated vs unauthenticated states. This also abstracts
the application coordinator further to manage both unauthenticated and
authenticated flows through the app. Since our target app requires
authentication for all functionality (outside of things like forgot
username/password, etc.), this structure will help manage the root of
the application and make it easier to swap log in and main scenes
easily.